+++ bits/bitaux/lib.rs
@@ -121,6 +121,12 @@ impl<T: Block> BitAux<Vec<T>> {
     }
 }
 
+impl<T> BitAux<T> {
+    pub fn inner(&self) -> &T {
+        &self.inner
+    }
+}
+
 impl<T: Unpack + Bits> Bits for BitAux<T> {
     #[inline]
     fn bits(&self) -> usize {
@@ -449,60 +455,3 @@ impl Poppy {
 //         Rho(Imp { buckets: buckets.into(), samples: None, bit_vec: dat.into() })
 //     }
 // }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use quickcheck::quickcheck;
-
-    fn none<T: Block>(n: usize) -> BitAux<Vec<T>> {
-        BitAux::new(n)
-    }
-
-    fn setup_bits(size: usize, mut bits: Vec<usize>) -> Vec<usize> {
-        bits.push(0);
-        bits.push((1 << 16) - 1);
-        bits.push(1 << 16);
-        bits.push(1 << 20);
-        bits.push(1 << 32);
-
-        let mut bits = bits.into_iter().filter(|&x| x < size).collect::<Vec<_>>();
-        bits.sort();
-        bits.dedup();
-        bits
-    }
-
-    fn check<T: Block + Unpack>(size: usize, bits: Vec<usize>) -> bool {
-        let mut aux = none::<T>(size);
-
-        for &b in &bits {
-            aux.bit_set(b);
-        }
-
-        assert_eq!(aux.count1(), bits.len());
-
-        bits.into_iter().enumerate().all(|(i, b)| {
-            aux.bit(b).unwrap()
-                && aux.rank1(..b) == i
-                && aux.select1(i) == Some(b)
-                && aux.inner.select1(i) == Some(b)
-        })
-    }
-
-    quickcheck! {
-        fn bits_u64(bits: Vec<usize>) -> bool {
-            let size = 1 << 18;
-            let bits = setup_bits(size, bits);
-
-            check::<u64>(size, bits)
-        }
-
-        fn bits_boxed_array(bits: Vec<usize>) -> bool {
-            let size = (1 << 32) + 65536;
-            let bits = setup_bits(size, bits);
-
-            check::<Box<[u64; 1024]>>(size, bits)
-        }
-    }
-}
