+++ bits/bitpacking/BUILD.bazel
@@ -4,10 +4,7 @@ package(default_visibility = ["//visibility:private"])
 
 rust_library(
     name = "bitpacking",
-    srcs = glob(
-        ["*.rs"],
-        exclude = ["*_test.rs"],
-    ),
+    srcs = ["bitpacking.rs"],
     visibility = ["//visibility:public"],
     deps = [
         "//bits",
@@ -15,30 +12,12 @@ rust_library(
     ],
 )
 
-rust_test_suite(
-    name = "tests",
-    srcs = glob(["*_test.rs"]),
-    proc_macro_deps = [
-        "@crates//:quickcheck_macros",
-    ],
-    deps = [
-        ":bitpacking",
-        "//bits",
-        "@crates//:quickcheck",
-    ],
-)
-
 rust_test(
-    name = "cfg_test",
+    name = "bitpacking_unit_test",
     crate = ":bitpacking",
 )
 
 rust_doc_test(
-    name = "doc_test",
-    crate = ":bitpacking",
-)
-
-rust_doc(
-    name = "doc",
+    name = "bitpacking_doc_test",
     crate = ":bitpacking",
 )
