+++ bits/tests/bits.rs
@@ -3,7 +3,9 @@ extern crate quickcheck_macros;
 
 use std::borrow::Cow;
 
-use bits::Bits;
+use bitaux::BitAux;
+use bitpacking::Unpack;
+use bits::{Bits, BitsMut, Block};
 
 #[test]
 fn bits_is_implemented() {
@@ -101,4 +103,56 @@ fn repr_select1(vec: Vec<u32>) -> bool {
 fn repr_select0(vec: Vec<u32>) -> bool {
     let aux = bitaux::BitAux::from(&vec[..]);
     (0..vec.count0()).all(|i| vec.select0(i) == aux.select0(i))
+}
+
+fn none<T: Block>(n: usize) -> BitAux<Vec<T>> {
+    BitAux::new(n)
+}
+
+fn setup_bits(size: usize, mut bits: Vec<usize>) -> Vec<usize> {
+    bits.push(0);
+    bits.push((1 << 16) - 512);
+    bits.push(1 << 16);
+    bits.push((1 << 16) + 512);
+    bits.push(1 << 20);
+    bits.push(1 << 32);
+    bits.push((1 << 32) + 65530);
+
+    let mut bits = bits.into_iter().filter(|&x| x < size).collect::<Vec<_>>();
+    bits.sort();
+    bits.dedup();
+    bits
+}
+
+fn check<T: Block + Unpack>(size: usize, bits: Vec<usize>) -> bool {
+    let mut aux = none::<T>(size);
+
+    for &b in &bits {
+        aux.bit_set(b);
+    }
+
+    assert_eq!(aux.count1(), bits.len());
+
+    bits.into_iter().enumerate().all(|(i, b)| {
+        aux.bit(b).unwrap()
+            && aux.rank1(..b) == i
+            && aux.select1(i) == Some(b)
+            && aux.inner().select1(i) == Some(b)
+    })
+}
+
+#[quickcheck]
+fn bits_u64(bits: Vec<usize>) -> bool {
+    let size = 1 << 18;
+    let bits = setup_bits(size, bits);
+
+    check::<u64>(size, bits)
+}
+
+#[quickcheck]
+fn bits_boxed_array(bits: Vec<usize>) -> bool {
+    let size = (1 << 32) + 65536;
+    let bits = setup_bits(size, bits);
+
+    check::<Box<[u64; 1024]>>(size, bits)
 }
